+++ app/assets/javascripts/util/ui_helper.js
@@ -36,43 +36,6 @@ var UIHelper = (function () {
     const RECENTS_METADATA = 'metadata';
 
     /*
-     * initStatedField - Initialize a Stated field radio button group
-     * @param [object or string] elementOrSelector - Either a jquery object or the class or ID selector (including the "#" or "." prefix) that represents the radio button group we are initializing.
-     * @param [string] startingValue - the value to set the initial state of the field to, options are 'true' or 'false'. Default is 'true'
-     * @param [function] onChangeFunction - a function object that will be run when the field is changed (does not take parameters)
-     */
-    function initStatedField(elementOrSelector, startingValue, onChangeFunction) {
-
-        var statedGroup;
-
-        // If the type of the first parameter is a string, then use it as a jquery selector, otherwise use as is
-        if (typeof elementOrSelector === "string") {
-            statedGroup = $(elementOrSelector);
-        } else {
-            statedGroup = elementOrSelector;
-        }
-
-        // if the starting value was not passed in then set it to the default
-        if (startingValue != null || startingValue != undefined){
-            startingValue = 'true';
-        }
-
-        // set the stated field value
-        setStatedField(statedGroup, startingValue);
-
-        // set the field change function to set the appropriate classes
-        statedGroup.change(function(){
-
-            statedGroup.parent().toggleClass("btn-primary btn-default");
-
-            // If it was passed run the supplied onChange function
-            if (onChangeFunction != null || onChangeFunction != undefined){
-                onChangeFunction();
-            }
-        });
-    }
-
-    /*
      * initDatePicker - Initialize a date picker input group to with a starting date
      * @param [object or string] elementOrSelector - Either a jquery object or the class or ID selector (including the "#" or "." prefix) that represents the date picker input group we are initializing.
      * @param [Number or string] startingDate - a long number that represents the date in milliseconds since the epoch, or the string 'latest'
@@ -120,74 +83,6 @@ var UIHelper = (function () {
     }
 
     /*
-     * initAllowedStatesField - Initialize an Allowed States field radio button group
-     * @param [object or string] elementOrSelector - Either a jquery object or the class or ID selector (including the "#" or "." prefix) that represents the radio button group we are initializing.
-     * @param [string] startingValue - the value to set the initial state of the field to, options are 'active', 'inactive', or 'active,inactive'. Default is 'active,inactive'
-     * @param [function] onChangeFunction - a function object that will be run when the field is changed (does not take parameters)
-     */
-    function initAllowedStatesField(elementOrSelector, startingValue, onChangeFunction) {
-
-        var allowedStatesGroup;
-
-        // If the type of the first parameter is a string, then use it as a jquery selector, otherwise use as is
-        if (typeof elementOrSelector === "string") {
-            allowedStatesGroup = $(elementOrSelector);
-        } else {
-            allowedStatesGroup = elementOrSelector;
-        }
-
-        // if the starting value was not passed in then set it to the default
-        if (startingValue != null || startingValue != undefined){
-            startingValue = 'active,inactive';
-        }
-
-        // set the stated field value
-        setAllowedStatesField(allowedStatesGroup, startingValue);
-
-        // set the field change function to set the appropriate classes
-        allowedStatesGroup.change(function(){
-
-            UIHelper.setAllowedStatesField(allowedStatesGroup, this.value);
-
-            // If it was passed run the supplied onChange function
-            if (onChangeFunction != null || onChangeFunction != undefined){
-                onChangeFunction();
-            }
-        });
-    }
-
-    /*
-     * setStatedField - set the value of a Stated field radio button group
-     * @param [object or string] elementOrSelector - Either a jquery object or the class or ID selector (including the "#" or "." prefix) that represents the radio button group whose value we are setting.
-     * @param [string] newValue - the value to set the state of the field to, options are 'true' or 'false'. Default is 'true'
-     */
-    function setStatedField(elementOrSelector, newValue) {
-
-        var statedGroup;
-
-        // If the type of the first parameter is a string, then use it as a jquery selector, otherwise use as is
-        if (typeof elementOrSelector === "string") {
-            statedGroup = $(elementOrSelector);
-        } else {
-            statedGroup = elementOrSelector;
-        }
-
-
-        // set the field value and classes by looping through each component and checking it's value against the passed in value
-        statedGroup.each(function (index, button) {
-
-            if (button.value == newValue) {
-
-                var buttonParent = button.parentElement;
-
-                button.checked = true;
-                buttonParent.classList.add('btn-primary');
-                $(buttonParent).removeClass("btn-default");
-            }
-        });
-    }
-
-    /*
      * setStampDate - set the value of a Stated field radio button group
      * @param [object or string] elementOrSelector - Either a jquery object or the class or ID selector (including the "#" or "." prefix) that represents the date picker input group whose value we are setting.
      * @param [string] newValue - the value to set the value of the field to
@@ -209,42 +104,6 @@ var UIHelper = (function () {
 
         // set the date input value
         dateGroup.find("input").val(newValue);
-    }
-
-    /*
-     * setAllowedStatesField - set the value of an Allowed States field radio button group
-     * @param [object or string] elementOrSelector - Either a jquery object or the class or ID selector (including the "#" or "." prefix) that represents the radio button group whose value we are setting.
-     * @param [string] newValue - the value to set the state of the field to, options are 'true' or 'false'. Default is 'true'
-     */
-    function setAllowedStatesField(elementOrSelector, newValue) {
-
-        var allowedStatesGroup;
-
-        // If the type of the first parameter is a string, then use it as a jquery selector, otherwise use as is
-        if (typeof elementOrSelector === "string") {
-            allowedStatesGroup = $(elementOrSelector);
-        } else {
-            allowedStatesGroup = elementOrSelector;
-        }
-
-
-        // set the field value and classes by looping through each component and checking it's value against the passed in value
-        allowedStatesGroup.each(function(index, button) {
-
-            var buttonParent = button.parentElement;
-
-            if (button.value == newValue) {
-
-                button.checked = true;
-                buttonParent.classList.add('btn-primary');
-                $(buttonParent).removeClass("btn-default");
-            } else {
-
-                button.checked = false;
-                buttonParent.classList.add("btn-default");
-                $(buttonParent).removeClass("btn-primary");
-            }
-        });
     }
 
     // function to the the position of an element from the edge of the viewpoint
@@ -1436,12 +1295,8 @@ var UIHelper = (function () {
     }
 
     return {
-        initStatedField: initStatedField,
         initDatePicker: initDatePicker,
-        initAllowedStatesField: initAllowedStatesField,
-        setStatedField: setStatedField,
         setStampDate: setStampDate,
-        setAllowedStatesField: setAllowedStatesField,
         getOffset: getOffset,
         getActiveTabId: getActiveTabId,
         isTabActive: isTabActive,