+++ ChangeLog.md
@@ -3,6 +3,15 @@ Komet Changelog
 This changelog summarizes changes and fixes which are a part of each revision.  For more details on the fixes, refer tracking numbers 
 where provided, and the git commit history.
 
+* 2017/03/15 - 1.67
+    * Refactored a number of the view params and added Module as a param (Jazz: 461738)
+    * Fixed Success messages to show in green banner, added success messages to concept create and edit, optimized some of the concept editing code
+    * Added status preference to taxonomy and concept viewer (Jazz: 461738)
+    * Changed status preference to only include Active and All (Jazz: 461738)
+    * Fixed concept create and edit to properly transfer view params to the tree after editing a concept (Jazz: 461738)
+    * Added path preference to view params for taxonomy, concepts, and mapping, though field is hidden currently  (Jazz: 461738)
+    * Added descriptive error messages if the two required fields for a mapset are not filled in. (Jazz: 476665)
+
 * 2017/03/08 - 1.66
     * Added role "alert" to error messages so they will be read by screen readers. They do not take the focus so that they don't interrupt the user flow, and because we add multiple at a time. Some messages can't contain detailed descriptions of the item affected if it is a new item. (Jazz: 469380)
     * Made concept viewer expand/collapse aria labels more descriptive of the sections they control. (Jazz: 469194)